requirements.txt
requests==2.32.3
PyYAML==6.0.2
python-docx==1.1.2
google-generativeai==0.8.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
